requirements.txt
google-adk>=0.1.0
google-genai>=0.6.0
google-cloud-logging>=3.10.0
python-dotenv>=1.0.1
requests>=2.32.0
langchain-community==0.3.20
wikipedia==1.4.0
flask>=3.0.0
pydantic>=2.8.0
